+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 
 * Integrate jointplot into pairplot, add point-estimate and overlay of plot kinds #1079
 ### Maintenance and fixes
+* Changed `diagonal` argument in `plot_pair` for `marginals` and fixed `point_estimate_marker_kwargs` in `plot_pair` 
 * Fixed behaviour of `credible_interval=None` in `plot_posterior` (#1115)
 * Fixed hist kind of `plot_dist` with multidimensional input (#1115)
 * Fixed `TypeError` in `transform` argument of `plot_density` and `plot_forest` when `InferenceData` is a list or tuple (#1121)